+++ projects/ng-infinite-scroll/src/public-api.ts
@@ -2,5 +2,6 @@
  * Public API Surface of ng-infinite-scroll
  */
 
+export * from './lib/component/ng-infinite-scroll.component';
+export * from './lib/directive/ng-infinite-scroll.directive';
 export * from './lib/ng-infinite-scroll.service';
-export * from './lib/ng-infinite-scroll.component';