+++ projects/ng-infinite-scroll/CHANGELOG.md
@@ -12,3 +12,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ci: basic setup for CI
 - chore: setup for Github and PRs
 - chore: basic library setup
+- feat: providing infinite scroll directive